Don't assume slt-fsp-pass-strength-result is set.
This fixes an issue we're having when users try to use your plugin with iTheme's Exchange. You shouldn't assume that your POST key is set as other plugins may access the WP function from elsewhere.

In our example, its being called via AJAX and the PHP warning is breaking the json for the callback.

Thanks!